fix(typescript): migrate storybook-doc Related to TypeScript
re #789
